A vIdeo processing software wIth the help of OOP, Opencv and GUI
I developed video processing software as a graduation project in the OOP course I took during my undergraduate education. In my project, I made a project that allows only one color range to be visible by changing the pixelization values on the image taken by the camera, converting from BGR to HSV format, thresholding, calculating the number of pixels in the final image, and performing them as a user from a single window.